myagents-py/requirements.txt
docker>=6.1.3
fastapi>=0.104.0
openai>=0.27.0
packaging>=23.2
pyautogen>=0.1.12
pydantic>=2.4.2
pydantic_core>=2.10.1
pyOpenSSL>=23.2.0
python-dotenv>=0.19.2
requests>=2.31.0
uvicorn>=0.23.2
websocket-client>=1.6.4
wsproto>=1.2.0
yarl>=1.9.2
fastapi>=0.104.0
pyautogen[retrievechat]
IPython